feat: add all methods, tests, and documentation
All proposed methods were implemented and tested.

The tests cover the happy path and occasionally the unhappy
path in range validation cases. This is done for both imperative
and recursive algorithms.

The unhappy path is usually discarded, as it deals with typing
validations, a native feature of Typescript.

A short description about the methods and how to run the tests
and the project is added to 'README.md'.